requirements.txt
apache-airflow==2.9.1
requests==2.31.0
pandas==2.2.2
pytest==8.2.2